requirements.txt
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.7
python-dotenv==1.0.1

SQLAlchemy==2.0.31
Werkzeug==3.0.3
alembic==1.13.2


